refact: make writer awaitable
Move ExperimentKind and experiment_kind to serialize.py
