+++ src/dipcoatimage/finitedepth/__init__.py
@@ -48,20 +48,12 @@ from .rectcoatinglayer import (
 from .experiment import ExperimentError, ExperimentBase, Experiment
 from .fastexperiment import FastExperiment
 from .analysis import (
-    ExperimentKind,
-    experiment_kind,
     AnalysisError,
     AnalysisBase,
     Analysis,
 )
 from .serialize import (
     data_converter,
-    ImportArgs,
-    ReferenceArgs,
-    SubstrateArgs,
-    CoatingLayerArgs,
-    ExperimentArgs,
-    AnalysisArgs,
     Config,
 )
 from .util import get_data_path
@@ -86,20 +78,10 @@ __all__ = [
     "ExperimentBase",
     "Experiment",
     "FastExperiment",
-    "ExperimentKind",
-    "experiment_kind",
-    "DataWriter",
-    "CSVWriter",
     "AnalysisError",
     "AnalysisBase",
     "Analysis",
     "data_converter",
-    "ImportArgs",
-    "ReferenceArgs",
-    "SubstrateArgs",
-    "CoatingLayerArgs",
-    "ExperimentArgs",
-    "AnalysisArgs",
     "Config",
     "get_data_path",
 ]
