+++ tests/test_analysis.py
@@ -1,45 +1,7 @@
 import csv
 import os
 import pytest
-from dipcoatimage.finitedepth import (
-    ExperimentKind,
-    experiment_kind,
-    get_data_path,
-)
 from dipcoatimage.finitedepth.analysis import CSVWriter
-
-
-def test_experiment_kind():
-    img_path = [
-        get_data_path("coat1.png"),
-    ]
-    assert experiment_kind(img_path) == ExperimentKind.SINGLE_IMAGE
-
-    imgs_path = [
-        get_data_path("coat1.png"),
-        get_data_path("coat2.png"),
-    ]
-    assert experiment_kind(imgs_path) == ExperimentKind.MULTI_IMAGE
-
-    vid_path = [
-        get_data_path("coat3.mp4"),
-    ]
-    assert experiment_kind(vid_path) == ExperimentKind.VIDEO
-
-    empty_path = []
-    assert experiment_kind(empty_path) == ExperimentKind.NULL
-    invalid_path = ["invalid.pdf"]
-    assert experiment_kind(invalid_path) == ExperimentKind.NULL
-    vids_path = [
-        get_data_path("coat3.mp4"),
-        get_data_path("coat3.mp4"),
-    ]
-    assert experiment_kind(vids_path) == ExperimentKind.NULL
-    vidimg_path = [
-        get_data_path("coat3.mp4"),
-        get_data_path("coat1.png"),
-    ]
-    assert experiment_kind(vidimg_path) == ExperimentKind.NULL
 
 
 def test_CSVWriter(tmp_path):
